refactored test_task2.py non exist test refactored task3.py - removed tests refactored test_task3.py to using parametrize refactored test_task3.py non exist test
