+++ hw5/task3.py
@@ -45,65 +45,3 @@ if __name__ == '__main__':
         print(f'There are no one who has salary lower than {thr}:')
 
     print(f'Salary average is {avr}')
-
-
-def test_non_exists_file():
-    assert ([], 0) == salary_analyze('1')
-
-
-def test_empty_file():
-    import os
-    f_path = 'tmp'
-    f = open(f_path, 'w')
-    f.close()
-    assert ([], 0) == salary_analyze(f_path)
-    os.remove(f_path)
-
-
-def test_positive():
-    import os
-    content = ['Ivanov:20000\n',
-               'Petrov:10000\n',
-               'Sidorov:15000\n',
-               'Medvedev:150000\n']
-    li = ['Petrov', 'Sidorov']
-    fs = 48750.0
-    f_path = 'tmp'
-    with open(f_path, 'w') as f:
-        f.writelines(content)
-
-    assert (li, fs) == salary_analyze(f_path)
-
-    os.remove(f_path)
-
-
-def test_empty_line():
-    import os
-    content = ['Ivanov:20000\n',
-               'Petrov:10000\n',
-               '\n',
-               'Medvedev:150000\n']
-    li = ['Petrov']
-    fs = 60000.0
-    f_path = 'tmp'
-    with open(f_path, 'w') as f:
-        f.writelines(content)
-
-    assert (li, fs) == salary_analyze(f_path)
-    os.remove(f_path)
-
-
-def test_not_a_number_value():
-    import os
-    content = ['Ivanov:20000\n',
-               'Petrov:10000\n',
-               'Sidorov:eee\n',
-               'Medvedev:150000\n']
-    li = ['Petrov']
-    fs = 60000.0
-    f_path = 'tmp'
-    with open(f_path, 'w') as f:
-        f.writelines(content)
-
-    assert (li, fs) == salary_analyze(f_path)
-    os.remove(f_path)
